Serve all content, even static content from a single WSGI handler
OpenShift, which we want to be able to deploy onto, is very limited in
it's handling of static content - only URLs that start with static/ can
be served statically. We can eventually migrate images, css, etc, there,
but for now, just serve static cotent directly out of the WSGI handler.

At a minimum, we should improve this by adding handling for If-Modified /
If-Match and/or by adding an expiration time.