Stop copying skeleton directories
Since we now are handling references to static files through our
WSGI handler rather than directly, it's no longer necessary to
copy over the skeleton directory for each planet - we can simply
fall back to the skeleton at lookup time.

We also use this approach to properly handle layering of dynamic
and static content in the output directory - with the OpenShift
layout with a separate output directory, this is no longer
automatic.