navbar.html: Temporarily unlink about.html
about.html has a lot of out-of-date information about planeteria.info; until
we figure out exactly how we are handling this site (do we strip it down to
just Planet Outreach and remove all planeteria content?) unlink about.htmt
from the header to not be confusing. We still link it from the footer since
there is some important legal information there.